+++ src/main/java/me/ajlane/geo/repast/colonisation/csr/package-info.java
@@ -1,0 +1,12 @@
+/**
+ * Classes used to implement a <emph>completely spatially random</emph> (CSR) colonisation model.
+ *
+ * <p>
+ * The model is completely spatially random in the sense that the locations of juvenile individuals
+ * distributed by the model are not correlated with the locations of land-cover patches containing
+ * corresponding mature vegetation.
+ * </p>
+ *
+ * @author Andrew Lane
+ */
+package me.ajlane.geo.repast.colonisation.csr;